hal: Initial dmac module with register wrappers
Defines a DMA controller (DMAC) module with volatile register wrappers.

Adds the following traits:

- `dmac::register::ReadableRegister`: registers with volatile reads
- `dmac::register::WritableRegister`: registers with volatile writes

Also adds a set of DMAC registers for both SAMD21 and SAMD51.

For SAMD51, defines `dmac::register::ChannelRegisters` which provides
a set of registers specific to each DMA channel, with each register
for an individual channel wrapped in a `VolatileCell`.

For now, the module is gated under the `unproven` feature. The goal is
to eventually use them to implement a DMA controller driver.